ISSUE #3776 - move all dispatchers together
